Removed timeouts, and re-added Resource download tests in separate unit test file
